Improved compatibility with new bot mods
Previously, Enemy.gd used the $ shorthand operator to get a node in host scene files. This could potentially cause issues if any host tried to spawn whose scene file did not include a node with that name. Now, the node is obtained with the get_node_or_null() method, and the update_look_direction method checks whether the node is present before updating.